+++ lib/string/shorten.ts
@@ -14,10 +14,10 @@ export default function shorten (val:string, length:number, postfix:string='...'
     if (typeof val !== 'string') return '';
 
     //  Return original value if options are invalid
-    if (typeof postfix !== 'string' || !Number.isFinite(length) || length <= 0) return val;
+    if (typeof postfix !== 'string' || !Number.isInteger(length) || length <= 0) return val;
 
     //  Trim first
     const sanitized = val.trim();
 
-    return sanitized.length <= length ? sanitized : `${sanitized.substr(0, length)}${postfix}`;
+    return sanitized.length <= length ? sanitized : `${sanitized.substring(0, length)}${postfix}`;
 }